Remove the overlay in the dashboard
this overlay was there to prevent scrolling and give the user a larger
click target. Now that scrolling is allowed there is no need for the
overlay.

The stars in the dashboard were in the overlay so they are now moved to
the FourUpComponent.

PT-175564402
