--- src/components/four-up-overlay.sass
@@ -1,17 +1,0 @@
-.four-up-overlay
-  position: absolute
-
-  .icon-holder
-    position: absolute
-    right: 10px
-    top: 35%
-
-    .icon-star
-      height: 42px
-      width: 42px
-      fill: #fff
-      opacity: 0.5
-      &.starred
-        fill: #fef171
-        opacity: 1
-
