--- src/components/four-up-overlay.tsx
@@ -1,66 +1,0 @@
-import { inject, observer } from "mobx-react";
-import React from "react";
-import { BaseComponent, IBaseProps } from "./base";
-import { DocumentViewMode } from "./document/document";
-import { DocumentModelType } from "../models/document/document";
-
-import "./four-up-overlay.sass";
-
-interface IProps extends IBaseProps {
-  style: any;
-  onClick: () => void;
-  documentViewMode?: DocumentViewMode;
-  document?: DocumentModelType;
-}
-
-@inject("stores")
-@observer
-export class FourUpOverlayComponent extends BaseComponent<IProps> {
-  public render() {
-    return (
-      <div
-        className="four-up-overlay"
-        style={this.props.style}
-        onClick={this.handleOverlayClick}
-      >
-        {this.renderStar()}
-      </div>
-    );
-  }
-
-  private renderStar() {
-    const { user } = this.stores;
-    const { documentViewMode, document } = this.props;
-    if (!document || (documentViewMode !== DocumentViewMode.Published)) {
-      return;
-    }
-
-    const isStarred = document.isStarredByUser(user.id);
-    return (
-      <div className="icon-holder" onClick={this.handleStarClick}>
-        <svg className={"icon-star " + (isStarred ? "starred" : "")} >
-          <use xlinkHref="#icon-outline-star"/>
-        </svg>
-      </div>
-    );
-  }
-
-  private handleOverlayClick = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (this.props.document) {
-      this.props.onClick();
-    }
-  };
-
-  private handleStarClick = (e: React.MouseEvent<HTMLDivElement>) => {
-    const {user} = this.stores;
-    const {document} = this.props;
-    e.preventDefault();
-    e.stopPropagation();
-    if (document) {
-      document.toggleUserStar(user.id);
-    }
-  };
-
-}
