+++ src/components/four-up.sass
@@ -124,3 +124,17 @@
       right: 0
       background-color: $id-purple
       border-bottom-left-radius: $border-radius
+
+  .icon-holder
+    position: absolute
+    right: 10px
+    top: 35%
+
+    .icon-star
+      height: 42px
+      width: 42px
+      fill: #fff
+      opacity: 0.5
+      &.starred
+        fill: #fef171
+        opacity: 1
